requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-CORS==3.1.13
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.20
python-dotenv==1.0.0
Werkzeug==2.3.2
